discord-bot/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
decorator==5.0.7
discord.py==1.7.1
ffmpeg-python==0.2.0
future==0.18.2
gTTS==2.2.2
idna==2.10
multidict==5.1.0
psycopg2-binary==2.8.6
pycparser==2.20
PyNaCl==1.4.0
python-slugify==4.0.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
text-unidecode==1.3
typing-extensions==3.7.4.3
urllib3==1.26.5
validators==0.18.2
yarl==1.6.3